requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
streamlit
Pillow
